enso_project

Documentation for enso_project.

enso_project.ESNHyperparams

enso_project.create_param_grid

enso_project.cross_validate_esn

enso_project.plot_prediction

enso_project.train_esn!

enso_project.ESNHyperparams — Type

Hyperparameters for an Echo State Network.

enso_project.create_param_grid — Method

create_param_grid(param_ranges:Array)

Given an Array of the parameter ranges, create a parameter grid.

enso_project.cross_validate_esn — Method

cross_validate_esn(train_data, val_data, param_grid)

Do a grid search on the given param_grid to find the optimal hyperparameters.

enso_project.plot_prediction — Method

plot_prediction(esn, Wₒᵤₜ, test_data)

Given an Echo State Network, plot its predictions versus the given test set.

enso_project.train_esn! — Method

train_esn!(esn, y, ridge_param)

Given an Echo State Network, train it on the target sequence y_target and return the optimised output weights Wₒᵤₜ.