enso_project

Documentation for enso_project. This Julia package provides tools to train Echo State Networks and Neural Differential Equations to predict dynamical systems.

enso_project.ChaoticNDE

enso_project.ESNHyperparams

enso_project.RandomSampler

enso_project.average_forecast_length

enso_project.create_param_grid

enso_project.cross_validate_esn

enso_project.esn_eval_pred

enso_project.forecast_lengths

enso_project.forecast_δ_1D

enso_project.plot_data_split_predictions

enso_project.plot_esn_prediction

enso_project.predict_node

enso_project.retrain_node

enso_project.setup_nn

enso_project.setup_node

enso_project.train_and_validate_node

enso_project.train_esn!

enso_project.train_node

enso_project.train_val_test_split

enso_project.ChaoticNDE — Type

ChaoticNDE{P,R,A,K} <: AbstractChaoticNDEModel

Model for setting up and training a Chaotic Neural Differential Equation.

Fields:

'p' parameter vector

'prob' NDEProblem

'alg' Algorithm to use for the 'solve' command

'kwargs' any additional keyword arguments that should be handed over (e.g. sensealg)

Constructor

'ChaoticNDE(prob; alg=Tsit5(), kwargs...)'

enso_project.ESNHyperparams — Type

Hyperparameters for an Echo State Network.

enso_project.RandomSampler — Type

RandomSampler(;kwargs...)

Sample randomly a hyperparameters configuration from the 'kwargs'.

enso_project.average_forecast_length — Method

average_forecast_length(predict, valid::NODEDataloader,N_t; λmax=0, mode="norm")

Returns the average forecast length on a NODEDataloader set (should be valid or test set) given a (t, u0) -> prediction function. N_t is the length of each forecast, has to be larger than the expected forecast length. If a λmax is given, the results are scaled with it (and dt`)

enso_project.create_param_grid — Method

create_param_grid(param_ranges:Array)

Given an Array of the parameter ranges, create a parameter grid.

enso_project.cross_validate_esn — Method

cross_validate_esn(train_data, val_data, param_grid)

Do a grid search on the given param_grid to find the optimal hyperparameters.

enso_project.esn_eval_pred — Method

esn_eval_pred(esn::ESN, W_out, data::Matrix)

given an ESN, its output layer W_out and a data matrix, evaluate the prediction of the ESN on the given data.

enso_project.forecast_lengths — Method

forecast_lengths(model, t::AbstractArray{T,1}, input_data::AbstractArray{T,S}, N_t::Integer=300; λ_max=0, mode="norm", threshold=0.4, output_data::Union{Nothing, AbstractArray{T,S}}=nothing)

Returns the forecast lengths of predictions on a NODEDataloader set (should be valid or test set) given a (t, u0) -> prediction function. N_t is the length of each forecast, has to be larger than the expected forecast length. If a λmax is given, the results are scaled with it (and dt`)

enso_project.forecast_δ_1D — Method

forecast_δ_1D(prediction::AbstractArray{T,N}, truth::AbstractArray{T,N}, mode::String="both") where {T,N}

Assumes that the last dimension of the input arrays is the time dimension and N_t long. Returns an N_t long array, judging how accurate the prediction is. Adapted such that it is possible to only consider the prediction's and truth's first dimension for calculation.

Supported modes:

"mean": mean between the arrays

"maximum": maximum norm

"norm": normalized, similar to the metric used in Pathak et al

"abs: absolute difference between arrays

enso_project.plot_data_split_predictions — Function

plot_data_split_predictions(predictions::Matrix, data::Matrix, splits::Array)

create one plot containing all subplots of different data splits.
Nrows corresponds to number of splits, Ncols is length of data to predict.
Splits: Array of integers corresponding to percentual splits

enso_project.plot_esn_prediction — Method

plot_esn_prediction(esn, Wₒᵤₜ, test_data, data_name::String)

Given an Echo State Network, plot its predictions versus the given test set. data_name is used to label the plot correctly

enso_project.predict_node — Method

predict_node(model, test::NODEData, data_name)

Given an optimal NDE 'model', return and plot its prediction on the given test set 'test'. 'data_name' is used to label the plot correctly.

enso_project.retrain_node — Method

retrain_node(training_data::NODEData, validation_data::NODEData, N_epochs, N_weights, N_hidden_layers, act, τ_max, η, seed)

Rerain the NDE with optimal hyperparameters 'Nweights', 'Nhiddenlayers', 'act', until integration length 'τmax' with learning rate 'η' on both the training and validation data together. For reproducibility of the results set random 'seed'.

enso_project.setup_nn — Method

setup_nn(N_weights, N_hidden_layers, act, seed, dimension)

Setup a NDE with input and output layers of 'dim' nodes, with weights 'Nweights', activation function 'act', number of hidden layers 'Nhidden_layers'. For reproducibility when choosing initial parameters for the network set random 'seed'.

enso_project.setup_node — Method

setup_node(pars, re_nn, u0, dt)

Setup NODE problem with parameters 'pars' of the neural network 're_nn' and intial value 'u0', time step 'dt'.

enso_project.train_and_validate_node — Method

train_and_validate_node(training_data::NODEData, validation_data::NODEData, N_epochs, N_weights, N_hidden_layers, act, τ_max, η, seed)

Sample randomly 'Nsamples' hyperparameter configs for the NDE and train the NDE for each config over 'Nepochs' epochs. For reproducibility of the results set random seed for each hyperparameter sample from given 'seeds'.

enso_project.train_esn! — Method

train_esn!(esn, y, ridge_param)

Given an Echo State Network, train it on the target sequence y_target and return the optimised output weights Wₒᵤₜ.

enso_project.train_node — Method

train_node(training_data::NODEData, validation_data::NODEData, N_epochs, N_weights, N_hidden_layers, act, τ_max, η, seed)

Train the NDE with weights 'Nweights', activation function 'act', until integration length 'τmax' with learning rate 'η'. For reproducibility of the results set random 'seed'.

enso_project.train_val_test_split — Method

train_val_test_split(data; num_val_months, num_test_months)

Split the given data into training, validation, and test sets. valpercent: number of time steps wanted in the validation set. testpercent: number of time steps wanted in the test set.